refactor(domain): restructure auditing components and improve folder organization
Refactored auditing logic by introducing separate folders for BaseClasses and Interfaces under the Auditing module.
This improves modularity and supports Screaming Architecture by clearly isolating responsibilities.

Also cleaned up Domain structure to enhance clarity and enforce separation of concerns, including clearer segregation for Entity, Enum, and Common layers.
