+++ TaskTrackerApp.Domain/Auditing/Interfaces/IAuditable.cs
@@ -1,0 +1,3 @@
+﻿namespace TaskTrackerApp.Domain.Auditing.Interfaces;
+
+public interface IAuditable : ICreationAuditable, IModificationAuditable { }
