+++ TaskTrackerApp.Domain/Auditing/Interfaces/ICreationAuditable.cs
@@ -1,0 +1,8 @@
+﻿namespace TaskTrackerApp.Domain.Auditing.Interfaces;
+
+public interface ICreationAuditable
+{
+    DateTime CreatedAt { get; set; }
+    string CreatedBy { get; set; }
+}
+
