+++ TaskTrackerApp.Domain/Auditing/Interfaces/IDeletionAuditable.cs
@@ -1,0 +1,8 @@
+﻿namespace TaskTrackerApp.Domain.Auditing.Interfaces;
+
+public interface IDeletionAuditable
+{
+    DateTime? DeletedAt { get; set; }
+    string? DeletedBy { get; set; }
+    bool IsDeleted { get; set; }
+}
