+++ TaskTrackerApp.Domain/Auditing/Interfaces/IFullAuditable.cs
@@ -1,0 +1,3 @@
+﻿namespace TaskTrackerApp.Domain.Auditing.Interfaces;
+
+public interface IFullAuditable : IAuditable, IDeletionAuditable { }
