+++ TaskTrackerApp.Domain/Auditing/Interfaces/IModificationAuditable.cs
@@ -1,0 +1,7 @@
+﻿namespace TaskTrackerApp.Domain.Auditing.Interfaces;
+
+public interface IModificationAuditable
+{
+    DateTime? UpdatedAt { get; set; }
+    string? UpdatedBy { get; set; }
+}
