+++ TaskTrackerApp.Domain/Common/BaseEntity.cs
@@ -1,0 +1,6 @@
+﻿namespace TaskTrackerApp.Domain.Common;
+
+public abstract class BaseEntity : IEntity
+{
+    public Guid Id { get; set; }
+}
